Add: source code for homework

+++ source_modified_model/geometry/node.hpp
@@ -1,43 +1,47 @@
 class Node {
     private:
-        short ID;
+        int ID;
         float x_coordinate;
         float y_coordinate;
         float z_coordinate;
 
     public:
-        Node(short identifier, float x_value, float y_value, float z_value){
+        Node(int identifier, float x_value, float y_value, float z_value) {
             ID = identifier;
             x_coordinate = x_value;
             y_coordinate = y_value;
             z_coordinate = z_value;
         }
 
-        void set_ID(short identifier){
+        void set_ID(int identifier) {
             ID = identifier;
         }
-        short get_ID(){
+
+        int get_ID() {
             return ID;
         }
 
-        void set_x_coordinate(float x_value){
+        void set_x_coordinate(float x_value) {
             x_coordinate = x_value;
         }
-        float get_x_coordinate(){
+
+        float get_x_coordinate() {
             return x_coordinate;
         }
 
-        void set_y_coordinate(float y_value){
+        void set_y_coordinate(float y_value) {
             y_coordinate = y_value;
         }
-        float get_y_coordinate(){
+
+        float get_y_coordinate() {
             return y_coordinate;
         }
 
-        void set_z_coordinate(float z_value){
+        void set_z_coordinate(float z_value) {
             z_coordinate = z_value;
         }
-        float get_z_coordinate(){
+
+        float get_z_coordinate() {
             return z_coordinate;
         }
 };
